awscdk-metaflow/requirements.txt
# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
aws-cdk-lib>=2.45.0, <3.0.0
constructs>=10.0.5, <11.0.0
pydantic
